Separated object printing task to another solution
