+++ ObjectPrintingTask/ObjectPrintingTask.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31729.503
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectPrintingTask", "ObjectPrintingTask\ObjectPrintingTask.csproj", "{A1E1DB53-D7B9-48D2-AA0D-B20ED621BAAB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectPrintingTests", "ObjectPrintingTests\ObjectPrintingTests.csproj", "{6550348E-79CE-4FB4-8387-61B6B0BB44CA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A1E1DB53-D7B9-48D2-AA0D-B20ED621BAAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1E1DB53-D7B9-48D2-AA0D-B20ED621BAAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1E1DB53-D7B9-48D2-AA0D-B20ED621BAAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1E1DB53-D7B9-48D2-AA0D-B20ED621BAAB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6550348E-79CE-4FB4-8387-61B6B0BB44CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6550348E-79CE-4FB4-8387-61B6B0BB44CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6550348E-79CE-4FB4-8387-61B6B0BB44CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6550348E-79CE-4FB4-8387-61B6B0BB44CA}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7BB63788-1213-4009-A74A-A8BBD4EABFDF}
+	EndGlobalSection
+EndGlobal
